Built MH Sampler, moved normalisation and tests
I built out the MetropolisHastingsSampler sampler class and used the test_sampler.cpp file to test the fits.
Added an extra cubic polynomial function that will be used to as a multidimensional example to show where the uniform sampler is outdone.
I added extra unit tests for more true rigidity setting in the observations class. Built out unit tests testing the constructor of the sampler class.
I moved the normalisation member function from the UniforrmSampler to the base abstract class as it is universal.