Wrote user instructions and app error handling
PREV COMMIT:

The previous commit that was made had an error where the commit message was not sent. The previous commit will be detailed below.

README.md was ammended to include few extra titles.

In MetropolisHastingsSampler.hpp file I include the assert library so that I could verify that at the end of every sample step the new_unit_hypercube and unit_hyper_cube (and the same of the parameter arrays) are the same to ensure updating is happening correectly.
Changed sample-points in set_extra_settings member function to N_sample and step_size to \sigma symbol. Pointless comments were deleted and the usage of bin_width in calculations was unnecessary as only the unit hypercube and param range is needed.

In Plot.hpp I added const and references to the function inputs and deleted commented out code for text on the graphs as step_size will remain constant in the application and number of samples can be include in the file name.

In Sampler.hpp I removed a this -> operator and replaced with . as it was in the base abstract class.

In the CMake files I added functionality to make a new Sample4D application.

In the UniformSamplingApp.cpp file I edited the help message to specify which parameters were optional. I also added functionality to switch plotting on and off.

I created the Sample4D app using the MixedSamplingApp.cpp file which had more parameters so more optional inputs and had the number of sampled input. I implemented factory method to chose between Samplers.

CURRENT COMMIT:

Filled in README.md file with instructions to run and buil Sample2D. Created .clang-format file but currently not functional. Changed \sigma_step back to step_size and symbols cannot be file names if zipping is used.
Added some spaces to .hpp files. I added more detail to the help messages in MixedSamplingApp.cpp and UniformSamplingApp.cpp. I added error handling to deal with incorrect range inputs being added.
In test_sampler.cpp I removed last MHS test case as no meaningful assertions were present.