Cleaned and documented committed
Added documentation for every class and function as neccessary. Removed meaningless assertions that do not depend on user input such as the fact that the sigma, input and output vectors of the Observation class template should be the same size.
Deleted useless comments and cleaned up error strings in the application files.